Add files via upload
Programming language used is “python”.

The software we used in this python programming are (requirements) : 

           1.pandas 

           2.numpy

           3.matplotlib

           4.scikit-learn

           5.pyQt5

In the back ground we have a offline collected data of patients in Excel format through which we have run this program according to the given data in it.  

The given data in it is: 1. Age, 2.sex, 3.bmi, 4.smoker, 5. Region.

 

The advantages 

           We would get a clear cut graphical view which would be use full to easily understand by the clint. 

(>>as  pictures speak more than words<<) 

 

So the outcome goes like this we would get the individual graph view of the given data.

 

 

 

 

\\A project by team “MAVERICK” @KIET HACKATHON NOVEMBER\\ 

ð M. NITYA SRI KRISHNA SAI (team leader)

ð B.VENKATESH KUMAR

ð A.VIJAYA LAKSHMI

ð M.DIVYA TRIDHARA

ð M. SAI LAKSHMI

ð V.DEVI

 













Thanking you iNeuron.ai Kakinada Institute of Engineering and Technology for your kind support to do this project by my own#hackathon2021#atkiet#emergingcourses#ineuronsupport#engineering#cse#ai&ds